+++ server/src/main/java/com/ncc/neon/better/PreprocessorNlpModule.java
@@ -22,7 +22,7 @@ public class PreprocessorNlpModule extends NlpModule {
     }
 
     @Override
-    protected void setEndpoints(HttpEndpoint[] endpoints) {
+    public void setEndpoints(HttpEndpoint[] endpoints) {
         for (HttpEndpoint endpoint : endpoints) {
             switch (endpoint.getType()) {
                 case PREPROCESS:
